Ported update tests
