+++ enginetest/engine_test.go
@@ -380,219 +380,6 @@ func TestDescribe(t *testing.T) {
 	})
 }
 
-func TestUpdate(t *testing.T) {
-	var updates = []struct {
-		updateQuery    string
-		expectedUpdate []sql.Row
-		selectQuery    string
-		expectedSelect []sql.Row
-	}{
-		{
-			"UPDATE mytable SET s = 'updated';",
-			[]sql.Row{{newUpdateResult(3,3)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "updated"}, {int64(2), "updated"}, {int64(3), "updated"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' WHERE i > 9999;",
-			[]sql.Row{{newUpdateResult(0,0)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' WHERE i = 1;",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "updated"}, {int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' WHERE i <> 9999;",
-			[]sql.Row{{newUpdateResult(3,3)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "updated"}, {int64(2), "updated"}, {int64(3), "updated"}},
-		},
-		{
-			"UPDATE floattable SET f32 = f32 + f32, f64 = f32 * f64 WHERE i = 2;",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM floattable WHERE i = 2;",
-			[]sql.Row{{int64(2), float32(3.0), float64(4.5)}},
-		},
-		{
-			"UPDATE floattable SET f32 = 5, f32 = 4 WHERE i = 1;",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT f32 FROM floattable WHERE i = 1;",
-			[]sql.Row{{float32(4.0)}},
-		},
-		{
-			"UPDATE mytable SET s = 'first row' WHERE i = 1;",
-			[]sql.Row{{newUpdateResult(1,0)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "second row"}, {int64(3), "third row"}},
-		},
-		{
-			"UPDATE niltable SET b = NULL WHERE f IS NULL;",
-			[]sql.Row{{newUpdateResult(2,1)}},
-			"SELECT * FROM niltable WHERE f IS NULL;",
-			[]sql.Row{{int64(4), nil, nil}, {nil, nil, nil}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' ORDER BY i ASC LIMIT 2;",
-			[]sql.Row{{newUpdateResult(2,2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "updated"}, {int64(2), "updated"}, {int64(3), "third row"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' ORDER BY i DESC LIMIT 2;",
-			[]sql.Row{{newUpdateResult(2,2)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "updated"}, {int64(3), "updated"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated' ORDER BY i LIMIT 1 OFFSET 1;",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "first row"}, {int64(2), "updated"}, {int64(3), "third row"}},
-		},
-		{
-			"UPDATE mytable SET s = 'updated';",
-			[]sql.Row{{newUpdateResult(3,3)}},
-			"SELECT * FROM mytable;",
-			[]sql.Row{{int64(1), "updated"}, {int64(2), "updated"}, {int64(3), "updated"}},
-		},
-		{
-			"UPDATE typestable SET ti = '2020-03-06 00:00:00';",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM typestable;",
-			[]sql.Row{{
-				int64(1),
-				int8(2),
-				int16(3),
-				int32(4),
-				int64(5),
-				uint8(6),
-				uint16(7),
-				uint32(8),
-				uint64(9),
-				float32(10),
-				float64(11),
-				sql.Timestamp.MustConvert("2020-03-06 00:00:00"),
-				sql.Date.MustConvert("2019-12-31"),
-				"fourteen",
-				false,
-				nil,
-				nil}},
-		},
-		{
-			"UPDATE typestable SET ti = '2020-03-06 00:00:00', da = '2020-03-06';",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM typestable;",
-			[]sql.Row{{
-				int64(1),
-				int8(2),
-				int16(3),
-				int32(4),
-				int64(5),
-				uint8(6),
-				uint16(7),
-				uint32(8),
-				uint64(9),
-				float32(10),
-				float64(11),
-				sql.Timestamp.MustConvert("2020-03-06 00:00:00"),
-				sql.Date.MustConvert("2020-03-06"),
-				"fourteen",
-				false,
-				nil,
-				nil}},
-		},
-		{
-			"UPDATE typestable SET da = '0000-00-00', ti = '0000-00-00 00:00:00';",
-			[]sql.Row{{newUpdateResult(1,1)}},
-			"SELECT * FROM typestable;",
-			[]sql.Row{{
-				int64(1),
-				int8(2),
-				int16(3),
-				int32(4),
-				int64(5),
-				uint8(6),
-				uint16(7),
-				uint32(8),
-				uint64(9),
-				float32(10),
-				float64(11),
-				sql.Timestamp.Zero(),
-				sql.Date.Zero(),
-				"fourteen",
-				false,
-				nil,
-				nil}},
-		},
-	}
-
-	for _, update := range updates {
-		e, idxReg := NewEngine(t)
-		ctx := enginetest.NewCtx(idxReg)
-		enginetest.TestQuery(t, ctx, e, update.updateQuery, update.expectedUpdate)
-		enginetest.TestQuery(t, ctx, e, update.selectQuery, update.expectedSelect)
-	}
-}
-
-func newUpdateResult(matched, updated int) sql.OkResult {
-	return sql.OkResult{
-		RowsAffected: uint64(updated),
-		Info:         plan.UpdateInfo{matched, updated, 0},
-	}
-}
-
-func TestUpdateErrors(t *testing.T) {
-	var expectedFailures = []struct {
-		name  string
-		query string
-	}{
-		{
-			"invalid table",
-			"UPDATE doesnotexist SET i = 0;",
-		},
-		{
-			"invalid column set",
-			"UPDATE mytable SET z = 0;",
-		},
-		{
-			"invalid column set value",
-			"UPDATE mytable SET i = z;",
-		},
-		{
-			"invalid column where",
-			"UPDATE mytable SET s = 'hi' WHERE z = 1;",
-		},
-		{
-			"invalid column order by",
-			"UPDATE mytable SET s = 'hi' ORDER BY z;",
-		},
-		{
-			"negative limit",
-			"UPDATE mytable SET s = 'hi' LIMIT -1;",
-		},
-		{
-			"negative offset",
-			"UPDATE mytable SET s = 'hi' LIMIT 1 OFFSET -1;",
-		},
-		{
-			"set null on non-nullable",
-			"UPDATE mytable SET s = NULL;",
-		},
-	}
-
-	for _, expectedFailure := range expectedFailures {
-		t.Run(expectedFailure.name, func(t *testing.T) {
-			e, idxReg := NewEngine(t)
-			_, _, err := e.Query(enginetest.NewCtx(idxReg), expectedFailure.query)
-			require.Error(t, err)
-		})
-	}
-}
-
 const testNumPartitions = 5
 
 func TestAmbiguousColumnResolution(t *testing.T) {
